TestJobs at feature level as Chaos Action config or Steps with events Bug fix in named parameter for Kriya Added plugin types to base configuration Moved random alphanumerical string generation from DataUtils to RandomizationUtils PnPRegistry handling null jar file and ignoring jars with missing plugin config file Fixed event names for Java test events

+++ framework/src/main/java/org/mvss/karta/framework/runtime/event/ScenarioCompleteEvent.java
@@ -25,10 +25,10 @@ public class ScenarioCompleteEvent extends Event
    private static final long serialVersionUID = 1L;
 
    private TestFeature       feature;
-   private long              iterationNumber;
+   private int               iterationNumber;
    private TestScenario      scenario;
 
-   public ScenarioCompleteEvent( String runName, TestFeature feature, long iterationNumber, TestScenario scenario )
+   public ScenarioCompleteEvent( String runName, TestFeature feature, int iterationNumber, TestScenario scenario )
    {
       super( StandardEventsTypes.SCENARIO_COMPLETE_EVENT, runName );
       this.feature = feature;
@@ -37,7 +37,7 @@ public class ScenarioCompleteEvent extends Event
    }
 
    @Builder
-   public ScenarioCompleteEvent( String runName, UUID id, Date timeOfOccurrence, TestFeature feature, long iterationNumber, TestScenario scenario )
+   public ScenarioCompleteEvent( String runName, UUID id, Date timeOfOccurrence, TestFeature feature, int iterationNumber, TestScenario scenario )
    {
       super( StandardEventsTypes.SCENARIO_COMPLETE_EVENT, runName, id, timeOfOccurrence );
       this.feature = feature;
