+++ framework/src/main/java/org/mvss/karta/framework/runtime/event/ScenarioStartEvent.java
@@ -25,10 +25,10 @@ public class ScenarioStartEvent extends Event
    private static final long serialVersionUID = 1L;
 
    private TestFeature       feature;
-   private long              iterationNumber;
+   private int               iterationNumber;
    private TestScenario      scenario;
 
-   public ScenarioStartEvent( String runName, TestFeature feature, long iterationNumber, TestScenario scenario )
+   public ScenarioStartEvent( String runName, TestFeature feature, int iterationNumber, TestScenario scenario )
    {
       super( StandardEventsTypes.SCENARIO_START_EVENT, runName );
       this.feature = feature;
@@ -37,7 +37,7 @@ public class ScenarioStartEvent extends Event
    }
 
    @Builder
-   public ScenarioStartEvent( String runName, UUID id, Date timeOfOccurrence, TestFeature feature, long iterationNumber, TestScenario scenario )
+   public ScenarioStartEvent( String runName, UUID id, Date timeOfOccurrence, TestFeature feature, int iterationNumber, TestScenario scenario )
    {
       super( StandardEventsTypes.SCENARIO_START_EVENT, runName, id, timeOfOccurrence );
       this.feature = feature;
